Added maxwidth, fixed weight bugs
'maxwidth' is a new command line parameter that gives a maximum bound on the
width of all dimensions. This is particularly useful for exploring hypercubes
(maxwidth=2).

There were some bugs that limited the viewable scope of the search algorithm to
networks where higher numbered dimensions had a weight of 1 weight=[2, 1]
even though the 'optimal' choice could be > 1 [2, 2]. An example of this is a
2D network making 32 nodes with radix 16 routers. You want Wi=[2,2] We=[4,4]
but the tool was reporting Wi=[2,4] We=[2,1].